# docs/openapi.json
{
  "components": {
    "schemas": {
      "Batch": {
        "additionalProperties": false,
        "properties": {
          "bytes": {
            "default": 16777216,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "rows": {
            "default": 10000,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "BatchList": {
        "properties": {
          "batches": {
            "items": {
              "properties": {
                "batch_id": {
                  "type": "integer"
                },
                "bytes": {
                  "type": "integer"
                },
                "detail": {
                  "description": "Serialized database receipt or error; no batch data",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "label": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "rows": {
                  "type": "integer"
                },
                "state": {
                  "enum": [
                    "INTENT",
                    "CONFIRMED",
                    "NOT_COMMITTED",
                    "UNKNOWN"
                  ]
                }
              },
              "required": [
                "batch_id",
                "state",
                "rows",
                "bytes",
                "label",
                "detail"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "batches"
        ],
        "type": "object"
      },
...
      "Consistency": {
        "enum": [
          "snapshot",
          "statement"
        ],
        "type": "string"
      },
      "Credentials": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "password_env": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string"
          },
          "commit_unknown": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "commit_unknown",
          "retryable"
        ],
        "type": "object"
      },
      "Execution": {
        "additionalProperties": false,
        "properties": {
          "bytes_per_second": {
            "default": null,
            "format": "uint64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "max_row_bytes": {
            "default": 25165824,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "memory_bytes": {
            "default": 268435456,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "queue_capacity": {
            "default": 4,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "rows_per_second": {
            "default": null,
            "format": "uint64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "timeout_ms": {
            "default": 3600000,
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "Field": {
        "properties": {
          "name": {
            "type": "string"
          },
          "nullable": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "nullable"
        ],
        "type": "object"
      },
...
      "Mapping": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "type": "object"
      },
      "MysqlConnection": {
        "additionalProperties": false,
        "properties": {
          "charset": {
            "default": "utf8mb4",
            "type": "string"
          },
          "credentials": {
            "$ref": "#/components/schemas/Credentials"
          },
          "database": {
            "type": "string"
          },
          "host": {
            "default": "127.0.0.1",
            "type": "string"
          },
          "max_allowed_packet_bytes": {
            "default": 67108864,
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "port": {
            "default": 3306,
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0,
            "type": "integer"
          },
          "session_variables": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "type": "object"
          },
          "time_zone": {
            "default": "+00:00",
            "type": "string"
          },
          "timeouts": {
            "$ref": "#/components/schemas/Timeouts",
            "default": {
              "connect_ms": 10000,
              "read_ms": 60000,
              "write_ms": 120000
            }
          },
          "tls": {
            "$ref": "#/components/schemas/Tls",
            "default": {
              "ca_certificate": null,
              "client_certificate": null,
              "client_key": null,
              "enabled": false,
              "server_name": null
            }
          }
        },
        "required": [
          "database",
          "credentials"
        ],
        "type": "object"
      },
...
      "ReaderConfig": {
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "batch": {
                "$ref": "#/components/schemas/Batch",
                "default": {
                  "bytes": 16777216,
                  "rows": 10000
                }
              },
              "connection": {
                "$ref": "#/components/schemas/MysqlConnection"
              },
              "consistency": {
                "$ref": "#/components/schemas/Consistency",
                "default": "snapshot"
              },
              "source": {
                "$ref": "#/components/schemas/Source"
              },
              "split": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Split"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": {
                "const": "mysql",
                "type": "string"
              }
            },
            "required": [
              "type",
              "connection",
              "source"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "batch": {
                "$ref": "#/components/schemas/Batch",
                "default": {
                  "bytes": 16777216,
                  "rows": 10000
                }
              },
              "credentials": {
                "$ref": "#/components/schemas/Credentials"
              },
              "database": {
                "type": "string"
              },
              "endpoint_map": {
                "additionalProperties": {
                  "type": "string"
                },
                "default": {},
                "type": "object"
              },
              "endpoint_parallelism": {
                "default": 1,
                "format": "uint",
                "minimum": 0,
                "type": "integer"
              },
              "flight_uri": {
                "type": "string"
              },
              "max_grpc_message_bytes": {
                "default": 67108864,
                "format": "uint",
                "minimum": 0,
                "type": "integer"
              },
              "session_variables": {
                "additionalProperties": {
                  "type": "string"
                },
                "default": {},
                "type": "object"
              },
              "source": {
                "$ref": "#/components/schemas/Source"
              },
              "timeouts": {
                "$ref": "#/components/schemas/Timeouts",
                "default": {
                  "connect_ms": 10000,
                  "read_ms": 60000,
                  "write_ms": 120000
                }
              },
              "tls": {
                "$ref": "#/components/schemas/Tls",
                "default": {
                  "ca_certificate": null,
                  "client_certificate": null,
                  "client_key": null,
                  "enabled": false,
                  "server_name": null
                }
              },
              "type": {
                "const": "doris",
                "type": "string"
              }
            },
            "required": [
              "type",
              "flight_uri",
              "database",
              "credentials",
              "source"
            ],
            "type": "object"
          }
        ]
      },
...
      "RequestStatus": {
        "properties": {
          "cancel_requested": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Run"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "request_id",
          "cancel_requested",
          "run"
        ],
        "type": "object"
      },
      "Run": {
        "properties": {
          "batches_committed": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "bytes_read": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "commit_unknown": {
            "type": "boolean"
          },
          "config": true,
          "created_at": {
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Error"
              },
              {
                "type": "null"
              }
            ]
          },
          "partial_write": {
            "type": "boolean"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "rows_committed": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "rows_filtered": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "rows_read": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "rows_submitted": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "run_id": {
            "type": "string"
          },
          "server_affected_rows": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "stage": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "state",
          "stage",
          "created_at",
          "updated_at",
          "rows_read",
          "bytes_read",
          "rows_submitted",
          "rows_committed",
          "rows_filtered",
          "server_affected_rows",
          "batches_committed",
          "partial_write",
          "commit_unknown",
          "config"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunSpec": {
        "additionalProperties": false,
        "properties": {
          "execution": {
            "$ref": "#/components/schemas/Execution",
            "default": {
              "bytes_per_second": null,
              "max_row_bytes": 25165824,
              "memory_bytes": 268435456,
              "queue_capacity": 4,
              "rows_per_second": null,
              "timeout_ms": 3600000
            }
          },
          "mapping": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Mapping"
            },
            "type": "array"
          },
          "reader": {
            "$ref": "#/components/schemas/ReaderConfig"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "writer": {
            "$ref": "#/components/schemas/WriterConfig"
          }
        },
        "required": [
          "reader",
          "writer"
        ],
        "title": "RunSpec",
        "type": "object"
      },
...
      "Validation": {
        "properties": {
          "semantics": {
            "type": "string"
          },
          "source_schema": {
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "type": "array"
          },
          "target_schema": {
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "type": "array"
          },
          "valid": {
            "const": true
          }
        },
        "required": [
          "valid",
          "source_schema",
          "target_schema",
          "semantics"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Arrow-native offline MySQL ↔ Doris. Batch commits; cancellation does not undo committed batches.",
    "title": "Dunnelean Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/requests/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            },
            "description": "Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/requests/{id}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            },
            "description": "Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/runs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 500,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "runs": {
                      "items": {
                        "$ref": "#/components/schemas/Run"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Run list"
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/runs/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/runs/{id}/batches": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchList"
                }
              }
            },
            "description": "Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/runs/{id}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/v1/validate": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            },
            "description": "Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:9876"
    }
  ]
}